Rewrite of day3
With this solution we get a runtime of 61ms.
The hashmap is doing a huge difference.
Explanation:

Prior we used an array and we searched for an element in
the array before we've added a new element. This took O(n)
in the worst case, the new algorithms looks up hash tables.
This works in constant time O(1).

Signed-off-by: Christian Rebischke <chris@nullday.de>